requirements.txt
django==1.11.17
django-braces
pillow
django-forms-bootstrap
django-kronos
invoiceGenerator